+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Include Link attributes (excluded id/class) in results 
+- Allow mulitple links using the same Url
 
 ## 1.0.0.0